pyproject.toml
[project]
name = "kakigori"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
license = { text = "AGPL-3.0-only" }
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.1.0",
    "brotli>=1.2.0",
    "cairosvg>=2.8.2",
    "fonttools>=4.61.1",
    "numpy>=2.0.0",
    "Pillow>=11.0.0",
    "svgpathtools>=1.7.2",
    "timm>=1.0.25",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.48.3",
    "tqdm>=4.67.3",
    "pypdfium2>=4.30.0",
    "verovio>=6.0.1",
    "torchmetrics>=1.8.2",
    "pycocotools>=2.0.11",
    "opencv-contrib-python>=4.13.0.92",
    "matplotlib>=3.10.8",
    "pyside6>=6.10.2",
    "albumentationsx>=2.0.18",
    "pyyaml>=6.0.3",
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "nbformat>=5.10.4",
    "music21>=9.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "ty>=0.0.18",
    "tensorboard>=2.19.0",
    "setuptools",
    "six",
    "ipykernel>=7.2.0",
    "tabulate>=0.9.0",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.uv.sources]
torch = [{ index = "pytorch-cu130" }]
torchvision = [{ index = "pytorch-cu130" }]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.ruff.lint]
select = ["I", "E402"]

[tool.ruff.lint.isort]
length-sort = true

[tool.ruff.lint.isort.import-heading]
future = "Future imports"
standard-library = "Standard library imports"
third-party = "Third party imports"
first-party = "First party imports"
local-folder = "Local folder imports"


[project.scripts]
render-dataset = "kakigori.dataset.render_dataset:main"
export-dataset = "kakigori.dataset.export_dataset:main"
generate-graphs = "kakigori.dataset.generate_graphs:main"
visualize-graphs = "kakigori.dataset.visualize_graphs:main"
extract-annotations = "kakigori.dataset.extract_annotations:main"
inject-mids = "kakigori.dataset.inject_mxl_ids:main"
extract-classes = "kakigori.dataset.extract_classes:main"
visualize-dataset = "kakigori.dataset.visualize_dataset:main"
synthetic-generator = "kakigori.dataset.synthetic_generator:main"
synthetic-writer = "kakigori.dataset.synthetic_writer:main"
filter-mxl = "kakigori.dataset.filter_mxl:main"
train-model = "kakigori.vision.train:main"
eval-model = "kakigori.vision.eval:main"
infer-model = "kakigori.vision.infer:main"
visualize-model = "kakigori.vision.visualize:main"
calculate-areas = "kakigori.vision.calculate_areas:main"
